Move info_row.dart to widget folder, create page title and shadow container components
