+++ lib/widgets/features/ShadowContainer.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class ShadowContainer extends StatelessWidget {
+  const ShadowContainer({
+    Key? key,
+    required this.title,
+    required this.content,
+  }) : super(key:key);
+
+  final String title;
+  final Widget content;
+
+  @override
+  Widget build(BuildContext context) {
+
+    throw UnimplementedError();
+  }
+}
