+++ lib/widgets/features/page_title.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+class PageTitle extends StatelessWidget {
+  const PageTitle ({
+    Key? key,
+    required this.title,
+  }) : super(key:key);
+
+  final String title;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: EdgeInsets.only(bottom: 40),
+      child: Text(title,
+          style: TextStyle(
+              color: Color(0xff2B8D78),
+              fontWeight: FontWeight.w700,
+              fontSize: 35
+          )),
+    );
+  }
+}
